requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile cookbook/llm_os/requirements.in
#
altair==5.3.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
appdirs==1.4.4
    # via yfinance
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via
    #   bs4
    #   yfinance
blinker==1.8.2
    # via streamlit
bs4==0.0.2
    # via -r cookbook/llm_os/requirements.in
cachetools==5.3.3
    # via streamlit
certifi==2024.2.2
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via curl-cffi
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   duckduckgo-search
    #   streamlit
    #   typer
contourpy==1.2.1
    # via matplotlib
curl-cffi==0.7.0b4
    # via duckduckgo-search
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via openai
duckdb==0.10.2
    # via -r cookbook/llm_os/requirements.in
duckduckgo-search==5.3.1
    # via -r cookbook/llm_os/requirements.in
exa-py==1.0.9
    # via -r cookbook/llm_os/requirements.in
fonttools==4.51.0
    # via matplotlib
frozendict==2.4.4
    # via yfinance
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via
    #   phidata
    #   streamlit
h11==0.14.0
    # via httpcore
html5lib==1.1
    # via yfinance
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   openai
    #   phidata
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
lxml==5.2.1
    # via yfinance
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.8.4
    # via -r cookbook/llm_os/requirements.in
mdurl==0.1.2
    # via markdown-it-py
multitasking==0.0.11
    # via yfinance
nest-asyncio==1.6.0
    # via -r cookbook/llm_os/requirements.in
numpy==1.26.4
    # via
    #   altair
    #   contourpy
    #   matplotlib
    #   pandas
    #   pgvector
    #   pyarrow
    #   pydeck
    #   streamlit
    #   yfinance
openai==1.28.1
    # via -r cookbook/llm_os/requirements.in
orjson==3.10.3
    # via duckduckgo-search
packaging==24.0
    # via
    #   altair
    #   matplotlib
    #   streamlit
pandas==2.2.2
    # via
    #   -r cookbook/llm_os/requirements.in
    #   altair
    #   streamlit
    #   yfinance
peewee==3.17.5
    # via yfinance
pgvector==0.2.5
    # via -r cookbook/llm_os/requirements.in
phidata==2.4.8
    # via -r cookbook/llm_os/requirements.in
pillow==10.3.0
    # via
    #   matplotlib
    #   streamlit
protobuf==4.25.3
    # via streamlit
psycopg[binary]==3.1.18
    # via -r cookbook/llm_os/requirements.in
psycopg-binary==3.1.18
    # via psycopg
pyarrow==16.0.0
    # via streamlit
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   openai
    #   phidata
    #   pydantic-settings
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
    # via phidata
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via matplotlib
pypdf==4.2.0
    # via -r cookbook/llm_os/requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via
    #   phidata
    #   pydantic-settings
pytz==2024.1
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.1
    # via phidata
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   exa-py
    #   streamlit
    #   yfinance
rich==13.7.1
    # via
    #   phidata
    #   streamlit
    #   typer
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   html5lib
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.30
    # via -r cookbook/llm_os/requirements.in
streamlit==1.34.0
    # via -r cookbook/llm_os/requirements.in
tenacity==8.3.0
    # via streamlit
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via phidata
toolz==0.12.1
    # via altair
tornado==6.4
    # via streamlit
tqdm==4.66.4
    # via openai
typer==0.12.3
    # via phidata
typing-extensions==4.11.0
    # via
    #   exa-py
    #   openai
    #   phidata
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
webencodings==0.5.1
    # via html5lib
yfinance==0.2.38
    # via -r cookbook/llm_os/requirements.in